Align methods in ProxyAnimation with the web-animations-2 spec (#23)
This patch provides a full polyfill implementation for the majority of the methods in the Animation interface, and stubs methods in the base interface (EventTarget). The ready and finished promise are polyfilled to closely match the specced behavior. Notable omissions are methods associated with replaceable animations, handling of hold phase, and direct handling of events (notably cancel and finish).

Authored-by: Kevin Ellis <kevers@chromium.org>