Rename isAnonymousScrollTimeline to isAnonymousTimeline
Although a ViewTimeline is also a ScrollTimeline, this rename is here to get consistency with the other functions.

All other functions use “Scroll” in their name when specifically targetting non-ViewTimeline ScrollTimelines, and have “View” in their name when targetting ViewTimeline ones.

As this function caters for both, the Scroll part is dropped from it.