Validate `axis` in constructor + fix axis update regression
Merging #140 introduced a regression making https://flackr.github.io/scroll-timeline/demo/view-timeline/ no longer work. Upon updating the axis always remained block.

This because the `this.axis` was directly maniuplated, instead of storing it in the options.

Furthermore, the setter – which validates the axis – didn’t get called in the constructor, so we need to expliclity validate the axis in the constructor as well.