Force parsing the linked stylesheet before animation starts
There is a chance for the animation to start before the polyfill
parses the linked stylesheet because we download and parse it
at the same time than the browser.

When this happens, the ScrollTimeline is not created because the
associated options are not available yet (
getAnimationTimelineOptions return null).

To prevent this from happening we remove the href of the linked
stylesheet to ensure that the polyfill will process the stylesheet
before the browser.